requirements.txt
torch==2.5.1
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torchvision==0.14.1
numpy==1.21.5
scikit-learn==1.0.2
Pillow==9.0.1